Got rid of nullspace monkeys

+++ code/modules/reagents/reagent_containers/food/mobcube.dm
@@ -4,14 +4,14 @@
 	icon_state = "monkeycube"
 	bitesize = 12
 	food_flags = FOOD_MEAT
-	var/mob/living/carbon/contained_mob = new /mob/living/carbon/monkey(null) //Storing it in nullspace is fine
+	var/contained_mob = /mob/living/carbon/monkey
 
 /obj/item/weapon/reagent_containers/food/snacks/monkeycube/New()
 	..()
 	reagents.add_reagent(NUTRIMENT,10)
 
 /obj/item/weapon/reagent_containers/food/snacks/monkeycube/Destroy()
-	if(contained_mob)
+	if(contained_mob && isdatum(contained_mob))
 		qdel(contained_mob)
 		contained_mob = null
 	..()
@@ -30,15 +30,23 @@
 
 /obj/item/weapon/reagent_containers/food/snacks/monkeycube/after_consume(var/mob/M)
 
+	if(!contained_mob)
+		return
+	if(ispath(contained_mob))
+		contained_mob = new contained_mob(src)
+	if(!ismob(contained_mob)) //Somehow
+		return
+	var/mob/C = contained_mob
+
 	to_chat(M, "<span class = 'warning'>Something inside of you suddently expands!</span>")
 
 	if (istype(M, /mob/living/carbon/human))
 		//Do not try to understand.
 		var/obj/item/weapon/surprise = new/obj/item/weapon(M)
-		surprise.icon = contained_mob.icon
-		surprise.icon_state = contained_mob.icon_state
-		surprise.name = "malformed [contained_mob.name]"
-		surprise.desc = "Looks like \a very deformed [contained_mob.name], a little small for its kind. It shows no signs of life."
+		surprise.icon = C.icon
+		surprise.icon_state = C.icon_state
+		surprise.name = "malformed [C.name]"
+		surprise.desc = "Looks like \a very deformed [C.name], a little small for its kind. It shows no signs of life."
 		qdel(contained_mob)
 		surprise.transform *= 0.6
 		surprise.add_blood(M)
@@ -56,10 +64,10 @@
 			E.embed(surprise)
 	else if (ismonkey(M))
 		M.visible_message("<span class='danger'>[M] suddenly tears in half!</span>")
-		contained_mob.forceMove(get_turf(M))
-		contained_mob.name = "malformed [initial(name)]"
-		contained_mob.transform *= 0.6
-		contained_mob.add_blood(M)
+		C.forceMove(get_turf(M))
+		C.name = "malformed [initial(name)]"
+		C.transform *= 0.6
+		C.add_blood(M)
 		M.gib()
 	..()
 
@@ -67,7 +75,13 @@
 
 	for(var/mob/M in viewers(src,7))
 		to_chat(M, "<span class='warning'>\The [src] expands!</span>")
-	contained_mob.forceMove(get_turf(src))
+	if(!contained_mob)
+		return
+	if(ispath(contained_mob))
+		new contained_mob(get_turf(src))
+	else if(ismob(contained_mob))
+		var/mob/C = contained_mob
+		C.forceMove(get_turf(src))
 	contained_mob = null
 	qdel(src)
 
@@ -87,27 +101,27 @@
 
 /obj/item/weapon/reagent_containers/food/snacks/monkeycube/farwacube
 	name = "farwa cube"
-	contained_mob = new /mob/living/carbon/monkey/tajara(null)
+	contained_mob = /mob/living/carbon/monkey/tajara
 
 /obj/item/weapon/reagent_containers/food/snacks/monkeycube/wrapped/farwacube
 	name = "farwa cube"
-	contained_mob = new /mob/living/carbon/monkey/tajara(null)
+	contained_mob = /mob/living/carbon/monkey/tajara
 
 /obj/item/weapon/reagent_containers/food/snacks/monkeycube/stokcube
 	name = "stok cube"
-	contained_mob = new /mob/living/carbon/monkey/unathi(null)
+	contained_mob = /mob/living/carbon/monkey/unathi
 
 /obj/item/weapon/reagent_containers/food/snacks/monkeycube/wrapped/stokcube
 	name = "stok cube"
-	contained_mob = new /mob/living/carbon/monkey/unathi(null)
+	contained_mob = /mob/living/carbon/monkey/unathi
 
 /obj/item/weapon/reagent_containers/food/snacks/monkeycube/neaeracube
 	name = "neaera cube"
-	contained_mob = new /mob/living/carbon/monkey/skrell(null)
+	contained_mob = /mob/living/carbon/monkey/skrell
 
 /obj/item/weapon/reagent_containers/food/snacks/monkeycube/wrapped/neaeracube
 	name = "neaera cube"
-	contained_mob = new /mob/living/carbon/monkey/skrell(null)
+	contained_mob = /mob/living/carbon/monkey/skrell
 
 /obj/item/weapon/reagent_containers/food/snacks/monkeycube/humancube
 	name = "humanoid cube"
